[FIX] mass_mailing_distribution_list: on mass mailings remove onchange on distribution_list_id
We can remove onchange as, when computing mailing_domain, we do NOT take care
of distribution_list_id if mailing_model_name is not distribution.list.
The problem we had with this onchange is that it automatically triggered
compute_mailing_domain at creation, which was unwanted in most cases
(for example it erased the mailing_domain given in context, if any).