[FIX] mozaik_membership_request: computing the price of the ML when partner was in a free state.
When partner was in a free state, the default Mozaik behavior was to create the following ML
with a null price. This caused problems for e.g. for supporters becoming members.
There was a first fix to trigger the product onchange to get the price of the product
on the ML before creating it. But this fix wasn't completely working: when the
price was overwritten on the MR, we forgot to update it in the ML creation vals.
This fix thus also deals with this special case.

refs #67035