[REF] Free memberships - refactoring: Before this commit the notion of free memberships was mixed with the notion of sponsored membership. These are two different notions. After this commit, it will be possible to: - Offer free memberships (sponsored or not) - Offer a sponsorship plan (with free or paying memberships)
We add on the membership requests an important notion of 'force subscription product'.
If a subscription product is forced, this one will be used on the membership line.
Otherwise the product will be determined with the amount specified on the membership line
and/or following the default subscription product of the partner.

refs #83285